:fire:  [FIX] 23/10/25 (09:14) 오류 FIX 및 옵션 추가, 챗봇 include
1. 모달창 출력 확인
 - 신청 내역 확인 시 목록의 사진 클릭하면 해당하는 사진을 확인할 수 있음
2. 채널 톡 api 스크립트 변경
3. 물품 선택 페이지 오류 수정
 - 카테고리 선택 시 버튼 색 변경
 - 물품 선택 창에선 규격을 변경하여도 잘 추가되었지만 사진 추가 페이지에서 선택한 정보가 아닌 option에서 기본으로 보여지는 정보로 들어가는 오류 수정
 4. 이미지 업로드 페이지 오류 FIX
  - 버튼 눌러도 input file이 클릭되지 않던 오류 수정
 5. 결제 페이지 css 수정
 6. 결제 완료 후 마이페이지로 이동 링크 연결
 7. 결제 취소 창에 selec로 취소 사유 선택 추가

+++ src/main/java/com/bringbring/reservation/domain/Pay.java
@@ -17,7 +17,7 @@ public class Pay {
     private String userId;
     private String payId;
     private String payMethod;
-    private double payAmount;
+    private long payAmount;
     private String payCurrency;
     private Date payDate;
     private char isPayStatus;
